move pf write and add resource writing
