+++ pkg/internal/tests/cross-tests/tfwrite.go
@@ -21,7 +21,6 @@ import (
 	"sort"
 
 	"github.com/hashicorp/hcl/v2/hclwrite"
-	pfproviderschema "github.com/hashicorp/terraform-plugin-framework/provider/schema"
 	"github.com/hashicorp/terraform-plugin-sdk/v2/helper/schema"
 	"github.com/pulumi/pulumi/sdk/v3/go/common/util/contract"
 	"github.com/zclconf/go-cty/cty"
@@ -59,95 +58,6 @@ func (w SDKv2Writer) Resource(
 	writeBlock(block.Body(), sch, config.AsValueMap())
 	_, err := f.WriteTo(w.out)
 	return err
-}
-
-type PFWriter struct{ out io.Writer }
-
-func WritePF(out io.Writer) PFWriter { return PFWriter{out} }
-
-func (w PFWriter) Provider(sch pfproviderschema.Schema, providerName string, config map[string]cty.Value) error {
-	if !cty.ObjectVal(config).IsWhollyKnown() {
-		return fmt.Errorf("WriteHCL cannot yet write unknowns")
-	}
-	f := hclwrite.NewEmptyFile()
-	block := f.Body().AppendNewBlock("provider", []string{providerName})
-	writePfProvider(block.Body(), sch, config)
-	_, err := f.WriteTo(w.out)
-	return err
-}
-
-func writePfProvider(body *hclwrite.Body, schemas pfproviderschema.Schema, values map[string]cty.Value) {
-	writePfObject(body, pfproviderschema.NestedBlockObject{
-		Attributes: schemas.Attributes,
-		Blocks:     schemas.Blocks,
-	}, values)
-}
-
-// writePfBlock writes the values for a single schema block to parentBody.
-//
-// Because blocks can be repeated (ListNestedBlock and SetNestedBlock), writePfBlock
-// can write an arbitrary number of blocks.
-//
-// For example, writing a list would add two blocks to parentBody:
-//
-//	writePfBlock("key", parentBody, ListNestedBlock{count: int}, cty.Value([{count: 1}, {count: 2}]))
-//
-//	key {
-//	  count = 1
-//	}
-//	key {
-//	  count = 2
-//	}
-//
-// This is why writePfBlock is called with parentBody, instead of with the block body
-// already created (as with [writeBlock]).
-func writePfBlock(key string, parentBody *hclwrite.Body, schemas pfproviderschema.Block, value cty.Value) {
-	switch schemas := schemas.(type) {
-	case pfproviderschema.ListNestedBlock:
-		for _, v := range value.AsValueSlice() {
-			b := parentBody.AppendNewBlock(key, nil).Body()
-			writePfObject(b, schemas.NestedObject, v.AsValueMap())
-		}
-	case pfproviderschema.SetNestedBlock:
-		values := value.AsValueSet().Values()
-		for _, v := range values {
-			b := parentBody.AppendNewBlock(key, nil).Body()
-			writePfObject(b, schemas.NestedObject, v.AsValueMap())
-		}
-	case pfproviderschema.SingleNestedBlock:
-		body := parentBody.AppendNewBlock(key, nil).Body()
-
-		if value.IsNull() {
-			return
-		}
-
-		writePfObject(body, pfproviderschema.NestedBlockObject{
-			Attributes: schemas.Attributes,
-			Blocks:     schemas.Blocks,
-		}, value.AsValueMap())
-	default:
-		contract.Failf("Unknown block type: %T", schemas)
-	}
-}
-
-func writePfObject(body *hclwrite.Body, schemas pfproviderschema.NestedBlockObject, values map[string]cty.Value) {
-	keys := make([]string, 0, len(values))
-	for k := range values {
-		keys = append(keys, k)
-	}
-	sort.Strings(keys)
-
-	for _, key := range keys {
-		if _, ok := schemas.Attributes[key]; ok {
-			body.SetAttributeValue(key, values[key])
-			continue
-		}
-		if block, ok := schemas.Blocks[key]; ok {
-			writePfBlock(key, body, block, values[key])
-			continue
-		}
-		contract.Failf("Could not find a attr or block for value key %q", key)
-	}
 }
 
 func writeBlock(body *hclwrite.Body, schemas map[string]*schema.Schema, values map[string]cty.Value) {
